New condition for hasNullRelease
After this commit a repository has to have a release marked as latest
and with its tag named "null" to be reported with the state
"hasNullRelease". Before this commit it has been sufficient if a release
with the tagname "null" exists at all. The difference can be seen in the
package-openssl repository. Before this commit the repository is
reported with state "hasNullRelease". After this commit it is reported
as "failure".

Signed-off-by: Malte Münch <muench@b1-systems.de>